# SYCL : Improved implementation of dp4a for the Nvidia backend