… ggml/src/ggml-sycl/dpct/blas.hpp
… ggml/src/ggml-sycl/dpct/device.hpp
… ggml/src/ggml-sycl/dpct/memory.hpp