- [x] I have read the [contributing guidelines](https://github.com/ggerganov/llama.cpp/blob/master/CONTRIBUTING.md)
- Self-reported review complexity:
  - [x] Low
  - [ ] Medium
  - [ ] High

---

This PR updates the helper headers to a newer version to make use of the dp4a ptx intrinsics.
Changes don't affect current performance as the SYCL mul mat dispatcher is currently avoiding the mmq/mmvq paths.

To test the performance of these changes, I re-enabled mmvq, similar to how the CUDA backend does.

```cpp
    use_dequantize_mul_mat_vec = use_dequantize_mul_mat_vec && !use_mul_mat_vec_q;
```

**Note**: I am not submitting changes to the dispatch as doing that affects negatively the performance on Intel platforms. I will create a separate PR for that, in order to have a proper discussion if needed without blocking this change.

---

Performance metrics:

No changes observed using `llama-bench`, both on an Intel(R) Data Center GPU Max 1100 and an Nvidia A100.
We currently have very poor performance on the host, so any gains here are hidden by other performance issues.

`test-backend-ops` using perf mode shows the following performance improvements:

| MUL_MAT                                                                 | old dp4a (GB/s) | new dp4a  (GB/s) | Relative Perf (Higher Better) |
|-------------------------------------------------------------------------|----------|----------|-------------------------------|
| MUL_MAT(type_a=q4_1,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])         | 1.21     | 1.85     | 1.53                          |
| MUL_MAT(type_a=q5_0,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])         | 1.47     | 1.92     | 1.31                          |
| MUL_MAT(type_a=q5_1,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])         | 1.50     | 1.88     | 1.25                          |
| MUL_MAT(type_a=q8_0,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])         | 1.49     | 2.09     | 1.40                          |
| MUL_MAT(type_a=q2_K,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])         | 1.29     | 1.70     | 1.32                          |
| MUL_MAT(type_a=q3_K,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])         | 1.40     | 1.77     | 1.26                          |
| MUL_MAT(type_a=q5_K,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])         | 1.41     | 1.88     | 1.33                          |
| MUL_MAT(type_a=q6_K,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])         | 1.50     | 1.92     | 1.28                          |
| MUL_MAT(type_a=iq2_xs,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])       | 1.36     | 1.69     | 1.24                          |
| MUL_MAT(type_a=iq2_s,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])        | 1.75     | 1.76     | 1.01                          |
| MUL_MAT(type_a=iq3_xxs,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])      | 1.76     | 1.76     | 1.00                          |
| MUL_MAT(type_a=iq1_s,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])        | 1.68     | 1.71     | 1.02                          |
| MUL_MAT(type_a=iq1_m,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])        | 1.70     | 1.70     | 1.00                          |
| MUL_MAT(type_a=iq4_nl,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])       | 1.83     | 1.83     | 1.00                          |
| MUL_MAT(type_a=iq3_s,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])        | 1.79     | 1.80     | 1.01                          |
| MUL_MAT(type_a=iq4_xs,type_b=f32,m=16,n=1,k=256,bs=[1,1],nr=[1,1])       | 1.84     | 1.85     | 1.01                          |