@Alcpz 2024-07-15
After some consideration, I think you are being reasonable wrt the inclusion of the asm. Thanks for the feedback.